AutoReproducer/references/paperbench/submissions/lbcs/submission/requirements.txt
torch>=1.13
torchvision>=0.14
numpy>=1.21
scipy>=1.8
pyyaml>=6.0
tqdm>=4.62
pillow>=9.0
